# PERF: Add BN_OPT_3 to allow enabling -O3 at the function level 